% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_data.R
\docType{data}
\name{chromstaRData}
\alias{chromstaRData}
\alias{euratrans}
\alias{rat}
\title{ChIP-seq data from the EURATRANS project}
\format{BED files with aligned reads.}
\source{
www.euratrans.eu
}
\description{
ChIP-seq data from the EURATRANS project for left-ventricle (lv) heart tissue in brown normay (BN) and spontaneous hypertensive rat (SHR). The data was downsampled to chr12 to reduce the file size for demonstration purposes.
}

