% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_data.R
\docType{data}
\name{experiment_table_H4K20me1}
\alias{experiment_table_H4K20me1}
\title{Experiment data table for EURATRANS H4K20me1-example}
\format{A data.frame with columns 'file', 'mark', 'condition', 'replicate' and 'pairedEndReads'.}
\description{
Experiment data table for the EURTRANS H4K20me1 data sets of left-ventricle (lv) heart tissue for usage in vignette examples of package \pkg{chromstaR}.
}

