% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_data.R
\docType{data}
\name{experiment_table_SHR}
\alias{experiment_table_SHR}
\title{Experiment data table for EURATRANS SHR-example}
\format{A data.frame with columns 'file', 'mark', 'condition', 'replicate' and 'pairedEndReads'.}
\description{
Experiment data table for the EURTRANS data sets of left-ventricle (lv) heart tissue in spontaneous hypertensive rat (SHR) for usage in vignette examples of package \pkg{chromstaR}.
}

