% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_data.R
\docType{data}
\name{expression_lv}
\alias{expression_lv}
\title{Expression data for the EURATRANS project}
\format{A data.frame with columns 'ensembl_gene_id', 'expression_BN' and 'expression_SHR'}
\description{
Expression values for left-ventricle (lv) heart tissue in brown norway (BN) and spontaneous hypertensive rat (SHR).
}

