% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_data.R
\docType{data}
\name{rn4_chrominfo}
\alias{rn4_chrominfo}
\title{Chromosome length information for rn4}
\format{A data.frame with chromosome and length information.}
\description{
Chromosome length information for rat assembly rn4.
}
\seealso{
\code{\link[GenomeInfoDb]{fetchExtendedChromInfoFromUCSC}}
}

